Now camera_calibration support existing photo dir.

+++ src/pupil_tracker/CameraCalibrator.cpp
@@ -17,19 +17,24 @@ using namespace boost::posix_time;
 
 int main (int argc, char *argv[])
 {
+    bool cameraFileExists;
     if (argc < 2) {
         cout<<argv[0]<<" cameraBaseDir"<<endl;
     }
     
     fs::path cameraBaseDir(argv[1]);
     string input;
-    cout<<"type your camera name:";
+    cout<<"type your camera name( if this named folder exists in the camera_calibration folder, the video capture step will be skipped):";
     cin >> input;
     string cameraname  = input;
     fs::path imagesDir = cameraBaseDir / input;
-    if (!fs::create_directories(imagesDir)) {
-        cout<<"cannot create dir. exit"<<endl;
-        return 0;
+    if (fs::exists(imagesDir)) {
+        cameraFileExists = true;
+    } else {
+        if (!fs::create_directories(imagesDir)) {
+            cout<<"cannot create dir. exit"<<endl;
+            return 0;
+        }
     }
     int gridw, gridh;
     float gridSize;
@@ -66,32 +71,33 @@ int main (int argc, char *argv[])
         }
     }
     
-    
-    
-    VideoCapture cam;
-    cam.open(0);
-    if(!cam.isOpened()){
-        cout<<"cannot open camera or video file."<<endl;
-        return 0;
-    }
-    
-    int imageCounter = 0;
-    Mat origin;
-    while(true){
-        bool success = captureImage(cam, origin,false);
-        imshow("Image Intake",origin);
-        
-        int c = waitKey(1)%256;
-        if(c == 'q')
-            break;
-        if (c == 32) {
-            fs::path filepath = imagesDir / string("img"+boost::lexical_cast<string>(imageCounter++)+".jpg");
-            imwrite(filepath.string(), origin);
+    if (cameraFileExists == false) {
+        VideoCapture cam;
+        cam.open(0);
+        if(!cam.isOpened()){
+            cout<<"cannot open camera or video file."<<endl;
+            return 0;
         }
-
+        
+        int imageCounter = 0;
+        Mat origin;
+        while(true){
+            captureImage(cam, origin,false);
+            imshow("Image Intake",origin);
+            
+            int c = waitKey(1)%256;
+            if(c == 'q')
+                break;
+            if (c == 32) {
+                fs::path filepath = imagesDir / string("img"+boost::lexical_cast<string>(imageCounter++)+".jpg");
+                imwrite(filepath.string(), origin);
+            }
+            
+        }
+        
     }
 
-    vector<fs::path> images = listFilesWithExtension(imagesDir, "img", "jpg");
+    vector<fs::path> images = listFilesWithExtension(imagesDir, "", "jpg");
     
     if (images.empty()) {
         cout<<"no image files."<<endl;
